python/requirements.txt
authlib~=1.3
python-dotenv~=1.0
httpx[http2]~=0.27
build~=1.1
aiofiles~=24.1
pydantic~=2.11.5
pytest-asyncio~=0.26.0
pytest~=8.3.5
pytest-httpx~=0.35
.